requirements.txt
-i https://pypi.org/simple
click==8.1.3
jinja2==3.1.2
markupsafe==2.1.3 ; python_version >= '3.7'
python-dotenv==1.0.0
